fix build

+++ src/utils/Vec.h
@@ -136,7 +136,7 @@ class Vec {
 
     T& operator[](int idx) const {
         CrashIf(idx < 0);
-        CrashIf(idx >= len);
+        CrashIf((size_t)idx >= len);
         return els[idx];
     }
